# Add max duration timeout