--- a/pkg/reconciler/revision/resources/deploy_test.go
+++ b/pkg/reconciler/revision/resources/deploy_test.go
@@ -111,7 +111,7 @@ var (
 			Name:  "REVISION_TIMEOUT_SECONDS",
 			Value: "45",
 		}, {
-			Name:  "REQUEST_MAX_DURATION_SECONDS",
+			Name:  "MAX_DURATION_SECONDS",
 			Value: "100",
 		}, {
 			Name: "SERVING_POD",
@@ -267,8 +267,8 @@ func defaultRevision() *v1.Revision {
 			UID: "1234",
 		},
 		Spec: v1.RevisionSpec{
-			TimeoutSeconds:            ptr.Int64(45),
-			RequestMaxDurationSeconds: ptr.Int64(100),
+			TimeoutSeconds:     ptr.Int64(45),
+			MaxDurationSeconds: ptr.Int64(100),
 		},
 	}
 }
@@ -420,8 +420,8 @@ func withContainers(containers []corev1.Container) RevisionOption {
 			PodSpec: corev1.PodSpec{
 				Containers: containers,
 			},
-			TimeoutSeconds:            ptr.Int64(45),
-			RequestMaxDurationSeconds: ptr.Int64(100),
+			TimeoutSeconds:     ptr.Int64(45),
+			MaxDurationSeconds: ptr.Int64(100),
 		}
 	}
 }