--- a/pkg/reconciler/revision/resources/queue.go
+++ b/pkg/reconciler/revision/resources/queue.go
@@ -194,8 +194,8 @@ func makeQueueContainer(rev *v1.Revision, cfg *config.Config) (*corev1.Container
 		ts = *rev.Spec.TimeoutSeconds
 	}
 	maxDurationTS := int64(0)
-	if rev.Spec.RequestMaxDurationSeconds != nil {
-		maxDurationTS = *rev.Spec.RequestMaxDurationSeconds
+	if rev.Spec.MaxDurationSeconds != nil {
+		maxDurationTS = *rev.Spec.MaxDurationSeconds
 	}
 	ports := queueNonServingPorts
 	if cfg.Observability.EnableProfiling {
@@ -279,7 +279,7 @@ func makeQueueContainer(rev *v1.Revision, cfg *config.Config) (*corev1.Container
 			Name:  "REVISION_TIMEOUT_SECONDS",
 			Value: strconv.Itoa(int(ts)),
 		}, {
-			Name:  "REQUEST_MAX_DURATION_SECONDS",
+			Name:  "MAX_DURATION_SECONDS",
 			Value: strconv.Itoa(int(maxDurationTS)),
 		}, {
 			Name: "SERVING_POD",