--- a/pkg/testing/v1/service.go
+++ b/pkg/testing/v1/service.go
@@ -182,10 +182,10 @@ func WithRevisionTimeoutSeconds(revisionTimeoutSeconds int64) ServiceOption {
 	}
 }
 
-// WithRevisionMaxDurationTimeoutSeconds sets revision max duration timeout
-func WithRevisionMaxDurationTimeoutSeconds(revisionMaxDurationTimeoutSeconds int64) ServiceOption {
+// WithRequestMaxDurationSeconds sets revision max duration timeout
+func WithRequestMaxDurationSeconds(requestMaxDurationSeconds int64) ServiceOption {
 	return func(service *v1.Service) {
-		service.Spec.Template.Spec.MaxDurationTimeoutSeconds = ptr.Int64(revisionMaxDurationTimeoutSeconds)
+		service.Spec.Template.Spec.RequestMaxDurationSeconds = ptr.Int64(requestMaxDurationSeconds)
 	}
 }