… config/core/300-resources/configuration.yaml
… config/core/300-resources/revision.yaml
… config/core/300-resources/service.yaml
… docs/serving-api.md
… pkg/reconciler/revision/resources/queue_test.go