Fixes #10851 

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

* Adds a timeout for max duration of a connection at the queue proxy side
* Exposes the timeout at the revision level without setting a default. See for more [here](https://github.com/knative/serving/pull/11988#discussion_r723315384).
*  Adds required unit tests, updates schema and docs.
* Adds a test in conformance suite.

Benchmark before and after for the http time handler:

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkTimeoutHandler/sequential-12     2014          2427          +20.51%
BenchmarkTimeoutHandler/parallel-12       557           702           +25.89%

benchmark                                 old allocs     new allocs     delta
BenchmarkTimeoutHandler/sequential-12     6              6              +0.00%
BenchmarkTimeoutHandler/parallel-12       6              6              +0.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkTimeoutHandler/sequential-12     686           711           +3.64%
BenchmarkTimeoutHandler/parallel-12       682           703           +3.08%
```