@skonto 2021-11-22
Failing test:
```
--- FAIL: TestAutoscaleSustaining (175.94s)
    --- FAIL: TestAutoscaleSustaining/aggregation-weightedExponential (49.47s)
        autoscale_test.go:109: Creating a new Route and Configuration
        service.go:113: Creating a new Service service autoscale-sustaining-aggregation-weighted-hxvxkzui
        crd.go:36:  resource {<nil> <nil> <*>{&TypeMeta{Kind:,APIVersion:,} &ObjectMeta{Name:autoscale-sustaining-aggregation-weighted-hxvxkzui,GenerateName:,Namespace:,SelfLink:,UID:,ResourceVersion:,Generation:0,CreationTimestamp:0001-01-01 00:00:00 +0000 UTC,DeletionTimestamp:<nil>,DeletionGracePeriodSeconds:nil,Labels:map[string]string{},Annotations:map[string]string{},OwnerReferences:[]OwnerReference{},Finalizers:[],ClusterName:,ManagedFields:[]ManagedFieldsEntry{},} 
```