@skonto 2021-11-22
@dprotaso @julz pls review. I will add a separate integration test once this is merged.