@skonto 2021-11-23
Added an e2e test, simplified timoutHandler logic.