@skonto 2021-12-02
Failures:

```
--- FAIL: TestGlobalResyncOnDefaultCMChange (0.16s)
    logger.go:130: 2021-11-24T12:21:50.828Z	ERROR	revision/controller.go:108	Failed to create resolver transport	{"error": "open /var/run/secrets/kubernetes.io/serviceaccount/ca.crt: no such file or directory"}
    logger.go:130: 2021-11-24T12:21:50.829Z	INFO	revision/controller.go:117	Fetch GitHub commit ID from kodata failed{error 26 0  "KO_DATA_PATH" does not exist or is empty}
```

```
2021/11/30 13:50:42 Error during command execution: unknown flag: --gateway-api-version
Step failed: ./test/e2e-tests.sh --gateway-api-version latest
```