@skonto 2021-12-06
@dprotaso 
> The http.TimeoutHandler adds a deadline to the http.Request but what if user-container doesn't respond to that and we send it another request?

Afaik the order of handler invocation is (assuming I understand the question correctly and to elaborate a bit): 

```
...
requestMetricsHandler -> NewTimeoutHandler -> ForwardedShimHandler 

-> ProxyHandler ->  requestAppMetricsHandler -> ConcurrencyStateHandler 

-> httpProxy  -> user container target
```

ProxyHandler(breaker) enforces the concurrency limits. `TimeoutHandler` propagates the cancellable ctx ([no timeout is assigned to it](https://github.com/knative/serving/blob/a794e1fdcbd6652d81af93a430afe94fdcf9abb0/pkg/http/handler/timeout.go#L67)) [h.handler.ServeHTTP(tw, r.WithContext(ctx))](https://github.com/knative/serving/blob/a794e1fdcbd6652d81af93a430afe94fdcf9abb0/pkg/http/handler/timeout.go#L100) so when it kills the request (defer), cancelling of the ctx should do proper ctx canceling as with any other failed request for all chained handlers. Could you describe more in detail the scenario you had in mind?