@dprotaso 2021-12-06
> Could you describe in detail the scenario you had in mind?

I think my concern isn't specific to the PR so it shouldn't gate it. But to elaborate - imagine a user container cc=1 that spends up to one minute processing the HTTP request prior to sending a response.

If we set the max duration to 30s the container could effectively be processing 2 requests - since we timed out the first one but then user container might not consider itself to be cancelled. This seems more of a runtime contract concern that needs to be fleshed out.