@skonto 2021-12-06
> If we set the max duration to 30s the container could effectively be processing 2 requests - since we timed out the first one but then user container might not consider itself to be cancelled.

I see. I guess it is up to the user to have proper logic in place to detect when http connection is closed eg. [node js](https://stackoverflow.com/questions/7348736/how-to-check-if-connection-was-aborted-in-node-js-server) or for some runtimes like go propagate context across api boundaries. From my pov the idea of using a timeout is to make sure we kill a connection, for Serving it means once this is done that the request is over. If user app is keep using resources it is badly written I guess.
What guarantees we provide is important of course and the contract should be such that users can react to a connection timeout. They cannot do much to avoid a request failure but they should be able to detect it to be able to keep up with load by terminating gracefully whatever operation they are doing. Timeouts are configured by the user at they end of the day (not imposed by us) and they are a rule for cancelling work asap.  We could of course add some good defaults to help UX.