@skonto 2021-12-07
```
Running github.com/client9/misspell with reviewdog 🐶 ...
  reviewdog: post failed for github.com/client9/misspell: fail to parse diff: Get "https://api.github.com/repos/knative/serving/pulls/12322": dial tcp 140.82.114.5:443: connect: connection timed out
  Error: Process completed with exit code 1.
```