@dprotaso 2021-12-08
> > If we set the max duration to 30s the container could effectively be processing 2 requests
> 
> Is this different from the existing first-byte timeout though? If the first-byte timeout times out a request but the user container continues processing it the same thing would happen, no?

it's not - just mentioning it for context

/lgtm
/approve
/hold

Holding in case anyone has any last minute bikeshedding on the property name otherwise LGTM