@julz 2021-12-10
lgtm, thanks @skonto, I'll let @dprotaso decide whether we want to unhold so close to release or wait till Tuesday to land (arguably it's a new field so the risk is probably pretty low I guess)